+++ TractorMod/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed "your pickaxe isn't strong enough" message shown near quarry boulder with a low-level pickaxe.
+* Fixed local seasonal tractor/garage textures no longer updated on season change after 4.21.0.
 
 ## 4.22.2
 Released 26 March 2025 for SMAPI 4.1.10 or later.